feat: rename to httpbinx
