--- httpbin/routers/inspection/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-from httpbin.routers.inspection import request as request_inspection
-from httpbin.routers.inspection import response as response_inspection
