+++ httpbinx/routers/__init__.py
@@ -2,17 +2,17 @@
 
 from fastapi import APIRouter
 
-from httpbin.routers import anything
-from httpbin.routers import auth
-from httpbin.routers import cookies
-from httpbin.routers import dynamicdata
-from httpbin.routers import httpmethods
-from httpbin.routers import images
-from httpbin.routers import redirects
-from httpbin.routers import responseformats
-from httpbin.routers import statuscodes
-from httpbin.routers.inspection import request_inspection
-from httpbin.routers.inspection import response_inspection
+from httpbinx.routers import anything
+from httpbinx.routers import auth
+from httpbinx.routers import cookies
+from httpbinx.routers import dynamicdata
+from httpbinx.routers import httpmethods
+from httpbinx.routers import images
+from httpbinx.routers import redirects
+from httpbinx.routers import responseformats
+from httpbinx.routers import statuscodes
+from httpbinx.routers.inspection import request_inspection
+from httpbinx.routers.inspection import response_inspection
 
 router = APIRouter()
 router.include_router(httpmethods.router, tags=['HTTP Methods'])
