+++ httpbinx/routers/anything.py
@@ -2,8 +2,8 @@
 from fastapi import APIRouter
 from starlette.requests import Request
 
-from httpbin.helpers import request_attrs_response
-from httpbin.schemas import RequestDictModel
+from httpbinx.helpers import request_attrs_response
+from httpbinx.schemas import RequestDictModel
 
 router = APIRouter()
