+++ httpbinx/routers/inspection/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from httpbinx.routers.inspection import request as request_inspection
+from httpbinx.routers.inspection import response as response_inspection
